fix(automation): consolidate workspace cleanup in watchdog
