+++ openhands/automation/watchdog.py
@@ -14,13 +14,22 @@ reset to bash-start + run budget + margin once the bash command starts
 The watchdog is mode-agnostic — all mode-specific logic is encapsulated
 in the ExecutionBackend (see automation/backends/).
 
-The same loop prunes ``integration_events``: it is the service's only periodic
-janitor, and a second loop buys nothing for one bounded DELETE.
+The same loop prunes ``integration_events`` and local-mode run workspaces: it
+is the service's only periodic janitor, and a second loop buys nothing for one
+bounded DELETE.
 """
 
 import asyncio
 import logging
-from datetime import datetime, timedelta
+import os
+import shutil
+from contextlib import suppress
+from dataclasses import dataclass
+from datetime import UTC, datetime, timedelta
+from enum import StrEnum
+from pathlib import Path
+from typing import Final
+from uuid import UUID
 
 from sqlalchemy import delete, inspect, select, update
 from sqlalchemy.engine import CursorResult
@@ -28,6 +37,7 @@ from sqlalchemy.ext.asyncio import AsyncSession, async_sessionmaker
 from sqlalchemy.orm import selectinload
 
 from openhands.automation.backends import get_backend
+from openhands.automation.backends.local import local_runs_root
 from openhands.automation.config import Settings, get_config
 from openhands.automation.models import (
     Automation,
@@ -56,6 +66,8 @@ logger = logging.getLogger("automation.watchdog")
 # a table the receive path writes to; the loop runs again a minute later, so a
 # backlog drains rather than being dropped.
 PRUNE_BATCH_SIZE = 5000
+WORKSPACE_PURGE_BATCH_SIZE: Final[int] = 50
+DELETE_ATTEMPT_FACTOR: Final[int] = 3
 
 
 async def _get_automation_keep_alive(
@@ -541,6 +553,311 @@ async def mark_stale_runs(
     return marked
 
 
+@dataclass(frozen=True, slots=True)
+class PurgeResult:
+    """Result of a workspace purge run."""
+
+    candidates_found: int
+    deleted: int
+    missing: int
+    refused: int
+    errors: int
+    bytes_freed: int
+
+
+class DeleteOutcome(StrEnum):
+    """Outcome of one bounded workspace deletion attempt."""
+
+    DELETED = "deleted"
+    MISSING = "missing"
+    REFUSED = "refused"
+    ERROR = "error"
+
+
+@dataclass(frozen=True, slots=True)
+class WorkspaceDeleteResult:
+    outcome: DeleteOutcome
+    bytes_freed: int
+
+
+def _empty_result(candidates_found: int = 0) -> PurgeResult:
+    return PurgeResult(candidates_found, 0, 0, 0, 0, 0)
+
+
+def _workspace_root(workspace_base: str | os.PathLike[str] | None) -> Path:
+    """Return the normalized root that owns all automation run directories."""
+    return local_runs_root(workspace_base).resolve(strict=False)
+
+
+def _workspace_path(
+    workspace_base: str | os.PathLike[str] | None, run_id: UUID
+) -> Path:
+    """Build a run path from a typed UUID, never from raw path input."""
+    return _workspace_root(workspace_base) / str(run_id)
+
+
+def _scan_candidates(runs_root: Path) -> dict[UUID, float]:
+    """Find direct, canonical UUID workspace directories under runs_root."""
+    if _is_link_or_junction(runs_root):
+        logger.warning("Refusing linked workspace root during scan: %s", runs_root)
+        return {}
+
+    try:
+        if not runs_root.is_dir():
+            return {}
+        resolved_root = runs_root.resolve(strict=True)
+        candidates: list[tuple[UUID, float]] = []
+        with os.scandir(runs_root) as entries:
+            for entry in entries:
+                with suppress(OSError, ValueError):
+                    candidate_path = Path(entry.path)
+                    if entry.is_symlink() or _is_link_or_junction(candidate_path):
+                        continue
+                    if not entry.is_dir(follow_symlinks=False):
+                        continue
+                    run_id = UUID(entry.name)
+                    if str(run_id) != entry.name:
+                        continue
+                    resolved_path = candidate_path.resolve(strict=True)
+                    if (
+                        resolved_path.parent != resolved_root
+                        or not resolved_path.is_dir()
+                    ):
+                        continue
+                    mtime = entry.stat(follow_symlinks=False).st_mtime
+                    candidates.append((run_id, mtime))
+    except FileNotFoundError:
+        return {}
+    except OSError as exc:
+        logger.warning("Failed to scan workspace root %s: %s", runs_root, exc)
+        return {}
+
+    candidates.sort(key=lambda item: (item[1], str(item[0])))
+    return dict(candidates)
+
+
+def _dir_size(path: Path) -> int:
+    """Best-effort reclaimed-space estimate; may undercount on races."""
+    total = 0
+    with suppress(OSError):
+        for dirpath, dirnames, filenames in os.walk(path, followlinks=False):
+            # Unlike symlinks, os.walk descends Windows junctions even with
+            # followlinks=False; prune them so external targets are not counted.
+            dirnames[:] = [
+                name
+                for name in dirnames
+                if not _is_link_or_junction(Path(dirpath) / name)
+            ]
+            for filename in filenames:
+                # Read the link itself rather than its target.
+                with suppress(OSError):
+                    total += (
+                        (Path(dirpath) / filename).stat(follow_symlinks=False).st_size
+                    )
+    return total
+
+
+def _is_link_or_junction(path: Path) -> bool:
+    """Return whether path is a symlink or Windows directory junction."""
+    return path.is_symlink() or path.is_junction()
+
+
+def _is_expired(timestamp: datetime | float, cutoff: datetime) -> bool:
+    """Return whether a trusted timestamp is older than cutoff."""
+    try:
+        if isinstance(timestamp, datetime):
+            return ensure_utc(timestamp) < cutoff
+        return datetime.fromtimestamp(timestamp, tz=UTC) < cutoff
+    except (OverflowError, OSError, ValueError):
+        return False
+
+
+def _is_terminal(status: AutomationRunStatus) -> bool:
+    return status in {
+        AutomationRunStatus.COMPLETED,
+        AutomationRunStatus.FAILED,
+        AutomationRunStatus.CANCELLED,
+        AutomationRunStatus.SKIPPED,
+    }
+
+
+def _delete_workspace(
+    workspace_base: str | os.PathLike[str] | None,
+    run_id: UUID,
+) -> WorkspaceDeleteResult:
+    """Delete one verified run directory without following root reparse points."""
+    runs_root = _workspace_root(workspace_base)
+    workspace_path = _workspace_path(workspace_base, run_id)
+
+    if _is_link_or_junction(runs_root):
+        logger.warning("Refusing linked workspace root: %s", runs_root)
+        return WorkspaceDeleteResult(DeleteOutcome.REFUSED, 0)
+    if _is_link_or_junction(workspace_path):
+        logger.warning("Refusing linked workspace path: %s", workspace_path)
+        return WorkspaceDeleteResult(DeleteOutcome.REFUSED, 0)
+    if not workspace_path.exists():
+        return WorkspaceDeleteResult(DeleteOutcome.MISSING, 0)
+
+    try:
+        resolved_root = runs_root.resolve(strict=True)
+        resolved_path = workspace_path.resolve(strict=True)
+    except FileNotFoundError:
+        return WorkspaceDeleteResult(DeleteOutcome.MISSING, 0)
+    except OSError as exc:
+        logger.warning("Failed to resolve workspace %s: %s", workspace_path, exc)
+        return WorkspaceDeleteResult(DeleteOutcome.ERROR, 0)
+
+    if resolved_path.parent != resolved_root or not resolved_path.is_dir():
+        logger.warning("Refusing workspace outside expected root: %s", workspace_path)
+        return WorkspaceDeleteResult(DeleteOutcome.REFUSED, 0)
+
+    size = _dir_size(workspace_path)
+    try:
+        shutil.rmtree(workspace_path)
+        return WorkspaceDeleteResult(DeleteOutcome.DELETED, size)
+    except FileNotFoundError:
+        return WorkspaceDeleteResult(DeleteOutcome.MISSING, 0)
+    except OSError as exc:
+        logger.warning("Failed to delete workspace %s: %s", workspace_path, exc)
+        return WorkspaceDeleteResult(DeleteOutcome.ERROR, 0)
+
+
+async def purge_terminal_workspaces(
+    session_factory: async_sessionmaker[AsyncSession],
+    workspace_base: str | os.PathLike[str] | None,
+    retention_seconds: int,
+    batch_size: int = WORKSPACE_PURGE_BATCH_SIZE,
+    shutdown_event: asyncio.Event | None = None,
+    failed_last_cycle: set[UUID] | None = None,
+) -> PurgeResult:
+    """Purge expired terminal and inactive orphan workspace directories.
+
+    The module batch constant bounds both the successful deletions and the
+    classification candidate set. Failed/refused attempts may use up to three
+    times that budget, so an undeletable directory cannot block newer workspaces
+    indefinitely. The watchdog passes failed_last_cycle to move those entries
+    behind fresh candidates on the next cycle.
+    """
+    if retention_seconds < 0:
+        raise ValueError("retention_seconds must be non-negative")
+    if batch_size <= 0:
+        raise ValueError("batch_size must be positive")
+    if retention_seconds == 0 or (
+        shutdown_event is not None and shutdown_event.is_set()
+    ):
+        return _empty_result()
+
+    cutoff = utcnow() - timedelta(seconds=retention_seconds)
+    candidates = await asyncio.to_thread(
+        _scan_candidates, _workspace_root(workspace_base)
+    )
+    if shutdown_event is not None and shutdown_event.is_set():
+        return _empty_result(len(candidates))
+
+    max_attempts = batch_size * DELETE_ATTEMPT_FACTOR
+    previous_failures = failed_last_cycle if failed_last_cycle is not None else set()
+    ordered_candidates = sorted(
+        candidates.items(),
+        key=lambda item: (
+            item[0] in previous_failures,
+            item[1],
+            str(item[0]),
+        ),
+    )[:max_attempts]
+
+    by_id: dict[UUID, tuple[AutomationRunStatus, datetime | None]] = {}
+    candidate_ids = [run_id for run_id, _root_mtime in ordered_candidates]
+    if candidate_ids:
+        # The single query is safe because the candidate list is bounded by
+        # the deletion-attempt budget (150 entries with the default settings).
+        async with session_factory() as session:
+            stmt = select(
+                AutomationRun.id,
+                AutomationRun.status,
+                AutomationRun.completed_at,
+            ).where(AutomationRun.id.in_(candidate_ids))
+            rows = (await session.execute(stmt)).all()
+            by_id.update(
+                (run_id, (status, completed_at))
+                for run_id, status, completed_at in rows
+            )
+
+    deleted = missing = refused = errors = bytes_freed = attempts = 0
+    failed_this_cycle: set[UUID] = set()
+    for run_id, root_mtime in ordered_candidates:
+        if deleted >= batch_size or attempts >= max_attempts:
+            break
+        if shutdown_event is not None and shutdown_event.is_set():
+            break
+
+        row = by_id.get(run_id)
+        if row is not None:
+            status, completed_at = row
+            if not _is_terminal(status):
+                continue
+        else:
+            completed_at = None
+
+        # Service-owned live runs retain an AutomationRun row. Orphans and
+        # terminal rows without completed_at use the scandir-captured root mtime
+        # as their inactivity guard.
+        age = completed_at if completed_at is not None else root_mtime
+        if not _is_expired(age, cutoff):
+            continue
+        if shutdown_event is not None and shutdown_event.is_set():
+            break
+
+        attempts += 1
+        delete_result = await asyncio.to_thread(
+            _delete_workspace, workspace_base, run_id
+        )
+        match delete_result.outcome:
+            case DeleteOutcome.MISSING:
+                missing += 1
+            case DeleteOutcome.REFUSED:
+                refused += 1
+                failed_this_cycle.add(run_id)
+            case DeleteOutcome.ERROR:
+                errors += 1
+                failed_this_cycle.add(run_id)
+            case DeleteOutcome.DELETED:
+                deleted += 1
+                bytes_freed += delete_result.bytes_freed
+                logger.info(
+                    "Purged workspace for run %s (%d bytes freed)",
+                    run_id,
+                    delete_result.bytes_freed,
+                )
+
+    if failed_last_cycle is not None:
+        failed_last_cycle.clear()
+        failed_last_cycle.update(failed_this_cycle)
+
+    result = PurgeResult(
+        len(candidates), deleted, missing, refused, errors, bytes_freed
+    )
+    if result.deleted:
+        logger.info(
+            "Purge complete: %d deleted, %d missing, %d refused, %d errors, "
+            "%d bytes freed (%d candidates scanned)",
+            result.deleted,
+            result.missing,
+            result.refused,
+            result.errors,
+            result.bytes_freed,
+            result.candidates_found,
+        )
+    else:
+        logger.debug(
+            "Purge no-op: %d missing, %d refused, %d errors (%d candidates scanned)",
+            result.missing,
+            result.refused,
+            result.errors,
+            result.candidates_found,
+        )
+    return result
+
+
 async def prune_integration_events(
     session_factory: async_sessionmaker[AsyncSession],
     settings: Settings,
@@ -581,6 +898,7 @@ async def watchdog_loop(
         "Watchdog started, scanning every %ds",
         interval,
     )
+    failed_workspace_deletions: set[UUID] = set()
 
     while True:
         if shutdown_event is not None and shutdown_event.is_set():
@@ -600,6 +918,17 @@ async def watchdog_loop(
                 logger.info("Pruned %d expired integration event(s)", pruned)
         except Exception:
             logger.exception("Error pruning integration events")
+
+        if settings.is_local_mode and settings.workspace_retention_seconds > 0:
+            try:
+                await purge_terminal_workspaces(
+                    session_factory=session_factory,
+                    workspace_base=settings.workspace_base,
+                    retention_seconds=settings.workspace_retention_seconds,
+                    failed_last_cycle=failed_workspace_deletions,
+                )
+            except Exception:
+                logger.exception("Error purging workspace directories")
 
         if shutdown_event is not None:
             try:
